@@ -2,7 +2,7 @@ package com.example.alexr.wiredbraincoffeeco;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 
 public class LandingActivity extends AppCompatActivity {
